loire/tone: Fix auth_id size and location.

+++ fpc_imp_loire_tone.c
@@ -42,7 +42,7 @@ typedef struct {
     struct QSEECom_handle *fpc_handle;
     struct qsee_handle_t* qsee_handle;
     struct qcom_km_ion_info_t ihandle;
-    uint32_t auth_id;
+    uint64_t auth_id;
 } fpc_data_t;
 
 static const char *fpc_error_str(int err)
